% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bound1}
\alias{bound1}
\title{Thailand Region 12 ADM1 Boundaries with Attributes}
\format{
An `sf` object with 7 features and the following attributes:
\describe{
  \item{Shape_Leng}{Length of the boundary in map units.}
  \item{Shape_Area}{Area of the boundary in map units.}
  \item{ADM1_EN}{Province name in English.}
  \item{ADM1_TH}{Province name in Thai.}
  \item{ADM1_PCODE}{Province administrative code.}
  \item{ADM0_EN}{Country name in English.}
  \item{ADM0_TH}{Country name in Thai.}
  \item{ADM0_PCODE}{Country administrative code.}
  \item{date}{Date the boundary was created.}
  \item{validOn}{Date the boundary became valid.}
  \item{geometry}{The geometry column containing spatial polygons.}
}
}
\source{
Data processed from Thailand's official administrative boundaries, available at
[Humanitarian Data Exchange (HDX)](https://data.humdata.org/dataset/thailand-administrative-boundaries).
}
\usage{
bound1
}
\description{
A dataset representing administrative boundaries (ADM1) of Region 12 in Thailand, along with key attributes.
}
\details{
This dataset provides detailed administrative information for Region 12 provinces in Thailand.
}
\examples{
\dontrun{
library(sf)
library(ggplot2)

bound1 # already lazy-loaded - no need to run data(bound1)

# Plot the ADM1 boundaries with province names
ggplot(bound1) +
  geom_sf(aes(fill = ADM1_EN)) +
  geom_sf_label(aes(label = ADM1_EN)) +
  ggtitle("Thailand Region 12 ADM1 Boundaries")
}
}
\keyword{datasets}
