% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decay_functions.R
\name{calc_gaussian_decay}
\alias{calc_gaussian_decay}
\title{Calculate Gaussian Decay Weights}
\usage{
calc_gaussian_decay(distance, sigma = 11.64884, ...)
}
\arguments{
\item{distance}{Numeric vector of distances}

\item{sigma}{Standard deviation parameter controlling spread}

\item{...}{Additional parameters (not used in gaussian decay)}
}
\value{
Numeric vector of weights
}
\description{
Calculate Gaussian Decay Weights
}
\keyword{internal}
