% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_functions.R
\name{calc_identity}
\alias{calc_identity}
\title{Identity weight function}
\usage{
calc_identity(x)
}
\arguments{
\item{x}{Numeric vector, matrix, or SpatRaster of weights}
}
\value{
Same object as input
}
\description{
Returns weights unchanged
}
\keyword{internal}
