% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decay_functions.R
\name{calc_power_decay}
\alias{calc_power_decay}
\title{Calculate Power Decay Weights}
\usage{
calc_power_decay(distance, sigma, ...)
}
\arguments{
\item{distance}{Numeric vector of distances}

\item{sigma}{Exponent parameter controlling rate of decay}

\item{...}{Additional parameters (not used in power decay)}
}
\value{
Numeric vector of weights
}
\description{
Calculate Power Decay Weights
}
\keyword{internal}
