% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-spax_diagnoses.R
\name{create_diagnostic_plots}
\alias{create_diagnostic_plots}
\title{Create standard diagnostic plots}
\usage{
create_diagnostic_plots(
  convergence_results = NULL,
  distribution_results = NULL
)
}
\arguments{
\item{convergence_results}{Results from analyze_convergence()}

\item{distribution_results}{Results from analyze_distribution()}
}
\value{
List of plot functions
}
\description{
Creates a collection of standard diagnostic plots for iteration analysis.
}
\keyword{internal}
