% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testnumplus.R
\name{create_personz}
\alias{create_personz}
\title{Create a personz object}
\usage{
create_personz(name, age)
}
\arguments{
\item{name}{Character string representing person's name}

\item{age}{Numeric value for person's age}
}
\value{
An object of class "personz"
}
\description{
Create a personz object
}
