% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-ae_atoms.R
\name{.ae_align_vector_to_frame}
\alias{.ae_align_vector_to_frame}
\title{Align vector values to a target frame by the vector field's full domain}
\usage{
.ae_align_vector_to_frame(vector_field, target_frame)
}
\description{
Align vector values to a target frame by the vector field's full domain
}
\keyword{internal}
