% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-ae_atoms.R
\name{.ae_is_scalar}
\alias{.ae_is_scalar}
\title{Is x a length-one scalar accepted by field Ops?}
\usage{
.ae_is_scalar(x)
}
\description{
Is x a length-one scalar accepted by field Ops?
}
\keyword{internal}
