% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-ae_atoms.R
\name{.ae_normalize_raster_by_cell}
\alias{.ae_normalize_raster_by_cell}
\title{Normalize raster layers by cell-retaining groups}
\usage{
.ae_normalize_raster_by_cell(field, by, method, a0 = 0)
}
\description{
Normalize raster layers by cell-retaining groups
}
\keyword{internal}
