% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-ae_atoms.R
\name{.ae_normalize_vector}
\alias{.ae_normalize_vector}
\title{Normalize vector fields by retained-axis groups}
\usage{
.ae_normalize_vector(field, by, method, a0 = 0)
}
\description{
Normalize vector fields by retained-axis groups
}
\keyword{internal}
