% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-ae_atoms.R
\name{.ae_ratio}
\alias{.ae_ratio}
\title{Pointwise ratio of two fields with zero-safe division}
\usage{
.ae_ratio(num, denom, zero = 0)
}
\description{
Pointwise ratio of two fields with zero-safe division
}
\keyword{internal}
