% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-ae_atoms.R
\name{.ae_sum_by_frame}
\alias{.ae_sum_by_frame}
\title{Sum numeric values by index-frame coordinate tuples}
\usage{
.ae_sum_by_frame(values, frame, by, na.rm = TRUE)
}
\description{
Sum numeric values by index-frame coordinate tuples
}
\keyword{internal}
