% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-validation.R
\name{.assert_class}
\alias{.assert_class}
\title{Assert input is of specific class(es)}
\usage{
.assert_class(x, classes, name = "Input")
}
\arguments{
\item{x}{Object to check}

\item{classes}{Character vector of valid classes}

\item{name}{Name of input for error messages}
}
\description{
Assert input is of specific class(es)
}
\keyword{internal}
