% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-validation.R
\name{.assert_cols_exist}
\alias{.assert_cols_exist}
\title{Assert columns exist in data frame}
\usage{
.assert_cols_exist(df, cols, df_name = "data.frame")
}
\arguments{
\item{df}{Data frame to check}

\item{cols}{Character vector of required column names}

\item{df_name}{Name of data frame for error messages}
}
\description{
Assert columns exist in data frame
}
\keyword{internal}
