% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-validation.R
\name{.assert_greater}
\alias{.assert_greater}
\title{Assert values are strictly greater than minimum}
\usage{
.assert_greater(x, min, name = "Value")
}
\arguments{
\item{x}{Numeric vector to check}

\item{min}{Minimum value to exceed}

\item{name}{Name of input for error messages}
}
\description{
Assert values are strictly greater than minimum
}
\keyword{internal}
