% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-validation.R
\name{.assert_greater_equal}
\alias{.assert_greater_equal}
\title{Assert values are greater than or equal to minimum}
\usage{
.assert_greater_equal(x, min, name = "Value")
}
\arguments{
\item{x}{Numeric vector to check}

\item{min}{Minimum allowed value}

\item{name}{Name of input for error messages}
}
\description{
Assert values are greater than or equal to minimum
}
\keyword{internal}
