% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-validation.R
\name{.assert_ifca}
\alias{.assert_ifca}
\title{Validate IFCA object for diagnostic functions}
\usage{
.assert_ifca(x, check_history = TRUE, name = "Input")
}
\arguments{
\item{x}{Object to validate}

\item{check_history}{Logical; whether to check for iteration history}

\item{name}{Name of input for error messages}
}
\description{
Validate IFCA object for diagnostic functions
}
\keyword{internal}
