% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-validation.R
\name{.assert_integer}
\alias{.assert_integer}
\title{Assert value is integer}
\usage{
.assert_integer(x, name = "Input")
}
\arguments{
\item{x}{Numeric vector to check}

\item{name}{Name of input for error messages}
}
\description{
Assert value is integer
}
\keyword{internal}
