% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-validation.R
\name{.assert_length}
\alias{.assert_length}
\title{Assert length matches expected value}
\usage{
.assert_length(len_x, expected, name = "Input")
}
\arguments{
\item{len_x}{Length to check}

\item{expected}{Expected length}

\item{name}{Name of input for error messages}
}
\description{
Assert length matches expected value
}
\keyword{internal}
