% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-validation.R
\name{.assert_lengths_match}
\alias{.assert_lengths_match}
\title{Assert two lengths match}
\usage{
.assert_lengths_match(
  len_x,
  len_y,
  name_x = "First input",
  name_y = "Second input"
)
}
\arguments{
\item{len_x}{First length to check}

\item{len_y}{Second length to check}

\item{name_x}{Name of first input for error messages}

\item{name_y}{Name of second input for error messages}
}
\description{
Assert two lengths match
}
\keyword{internal}
