% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-validation.R
\name{.assert_less}
\alias{.assert_less}
\title{Assert values are strictly less than maximum}
\usage{
.assert_less(x, max, name = "Value")
}
\arguments{
\item{x}{Numeric vector to check}

\item{max}{Maximum value to not exceed}

\item{name}{Name of input for error messages}
}
\description{
Assert values are strictly less than maximum
}
\keyword{internal}
