% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-validation.R
\name{.assert_less_equal}
\alias{.assert_less_equal}
\title{Assert values are less than or equal to maximum}
\usage{
.assert_less_equal(x, max, name = "Value")
}
\arguments{
\item{x}{Numeric vector to check}

\item{max}{Maximum allowed value}

\item{name}{Name of input for error messages}
}
\description{
Assert values are less than or equal to maximum
}
\keyword{internal}
