% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-validation.R
\name{.assert_no_na}
\alias{.assert_no_na}
\title{Assert no NA values present}
\usage{
.assert_no_na(x, name = "Input")
}
\arguments{
\item{x}{Object to check for NAs}

\item{name}{Name of input for error messages}
}
\description{
Assert no NA values present
}
\keyword{internal}
