% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-validation.R
\name{.assert_numeric}
\alias{.assert_numeric}
\title{Assert input is numeric}
\usage{
.assert_numeric(x, name = "Input")
}
\arguments{
\item{x}{Object to check}

\item{name}{Name of input for error messages}
}
\description{
Assert input is numeric
}
\keyword{internal}
