% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-validation.R
\name{.assert_positive}
\alias{.assert_positive}
\title{Assert values are positive}
\usage{
.assert_positive(x, allow_zero = FALSE, name = "Input")
}
\arguments{
\item{x}{Numeric vector to check}

\item{allow_zero}{Whether to include zero as valid value}

\item{name}{Name of input for error messages}
}
\description{
Assert values are positive
}
\keyword{internal}
