% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-validation.R
\name{.assert_raster}
\alias{.assert_raster}
\title{Assert input is a SpatRaster}
\usage{
.assert_raster(x, name = "Input")
}
\arguments{
\item{x}{Object to check}

\item{name}{Name of input for error messages}
}
\description{
Assert input is a SpatRaster
}
\keyword{internal}
