% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-validation.R
\name{.assert_raster_alignment}
\alias{.assert_raster_alignment}
\title{Assert two rasters have matching geometry}
\usage{
.assert_raster_alignment(
  x,
  y,
  name_x = "First raster",
  name_y = "Second raster"
)
}
\arguments{
\item{x}{First raster to check}

\item{y}{Second raster to check}

\item{name_x}{Name of first raster for error messages}

\item{name_y}{Name of second raster for error messages}
}
\description{
Assert two rasters have matching geometry
}
\keyword{internal}
