% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-calc_choice.R
\name{.calc_choice_core}
\alias{.calc_choice_core}
\title{Internal function for core choice computation}
\usage{
.calc_choice_core(weights, attractiveness, alpha, a0, snap)
}
\description{
Internal function for core choice computation
}
\keyword{internal}
