% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-spax_diag_convergence.R
\name{.calc_efficiency_metrics}
\alias{.calc_efficiency_metrics}
\title{Calculate efficiency-related convergence metrics}
\usage{
.calc_efficiency_metrics(history)
}
\arguments{
\item{history}{Array of iteration history}
}
\value{
List of efficiency metrics
}
\description{
Calculate efficiency-related convergence metrics
}
\keyword{internal}
