% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-spax_diag_facil.R
\name{.calc_facility_convergence}
\alias{.calc_facility_convergence}
\title{Calculate facility-specific convergence characteristics}
\usage{
.calc_facility_convergence(history)
}
\arguments{
\item{history}{Array of iteration history}
}
\value{
Data frame of convergence metrics by facility
}
\description{
Calculate facility-specific convergence characteristics
}
\keyword{internal}
