% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-spax_diag_facil.R
\name{.calc_facility_extended}
\alias{.calc_facility_extended}
\title{Calculate additional facility-specific metrics beyond basic convergence}
\usage{
.calc_facility_extended(history)
}
\arguments{
\item{history}{Array of iteration history}
}
\value{
Data frame of extended facility metrics
}
\description{
Calculate additional facility-specific metrics beyond basic convergence
}
\keyword{internal}
