% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-spax_diag_convergence.R
\name{.calc_speed_metrics}
\alias{.calc_speed_metrics}
\title{Calculate speed-related convergence metrics}
\usage{
.calc_speed_metrics(history)
}
\arguments{
\item{history}{Array of iteration history}
}
\value{
List of speed metrics
}
\description{
Calculate speed-related convergence metrics
}
\keyword{internal}
