% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-spax_diag_convergence.R
\name{.calc_stability_metrics}
\alias{.calc_stability_metrics}
\title{Calculate stability-related convergence metrics}
\usage{
.calc_stability_metrics(history)
}
\arguments{
\item{history}{Array of iteration history}
}
\value{
List of stability metrics
}
\description{
Calculate stability-related convergence metrics
}
\keyword{internal}
