% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-calc_choice.R
\name{.chck_calc_choice}
\alias{.chck_calc_choice}
\title{Internal function for core choice cvalidation}
\usage{
.chck_calc_choice(weights, attractiveness = NULL, alpha = 1, a0 = 0)
}
\description{
Internal function for core choice cvalidation
}
\keyword{internal}
