% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-validation.R
\name{.chck_class}
\alias{.chck_class}
\title{Check if input is of a specific structure* (or range of classes)}
\usage{
.chck_class(x, classes, name = "Input")
}
\description{
Check if input is of a specific structure* (or range of classes)
}
\keyword{internal}
