% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-validation.R
\name{.chck_cols_exist}
\alias{.chck_cols_exist}
\title{Check if columns exist in dataframe}
\usage{
.chck_cols_exist(df, cols, df_name = "data.frame")
}
\description{
Check if columns exist in dataframe
}
\keyword{internal}
