% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-decay_functions.R
\name{.chck_decay}
\alias{.chck_decay}
\title{Validate inputs for decay calculation}
\usage{
.chck_decay(distance, method, sigma = NULL, snap = FALSE)
}
\arguments{
\item{distance}{Numeric vector, matrix, or SpatRaster of distances}

\item{method}{Character string or function specifying decay method}

\item{sigma}{Numeric parameter controlling decay rate}

\item{snap}{Logical for validation skipping}
}
\description{
Validate inputs for decay calculation
}
\keyword{internal}
