% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-spax_e2sfca.R
\name{.chck_e2sfca}
\alias{.chck_e2sfca}
\title{Validate inputs for spax_e2sfca}
\usage{
.chck_e2sfca(
  demand,
  supply,
  distance,
  decay_params,
  demand_normalize,
  id_col = NULL,
  supply_cols = NULL
)
}
\arguments{
\item{demand}{SpatRaster representing spatial distribution of demand}

\item{supply}{vector, matrix, or data.frame containing supply capacity values}

\item{distance}{SpatRaster stack of travel times/distances to facilities}

\item{decay_params}{List of parameters for decay function}

\item{demand_normalize}{Character specifying normalization method}

\item{id_col}{Character; column name for facility IDs if supply is a data.frame}

\item{supply_cols}{Character vector; names of supply columns if supply is a data.frame}
}
\description{
Validate inputs for spax_e2sfca
}
\keyword{internal}
