% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-gather_functions.R
\name{.chck_gather_weighted}
\alias{.chck_gather_weighted}
\title{Validate inputs for gather_weighted}
\usage{
.chck_gather_weighted(values, weights, na.rm)
}
\arguments{
\item{values}{SpatRaster of values}

\item{weights}{SpatRaster of weights}

\item{na.rm}{Logical for NA handling}
}
\description{
Validate inputs for gather_weighted
}
\keyword{internal}
