% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-spax_diagnoses.R
\name{.chck_history}
\alias{.chck_history}
\title{Validate iteration history array}
\usage{
.chck_history(history, metric_names = NULL)
}
\arguments{
\item{history}{Array of iteration history values}

\item{metric_names}{Character vector of metric names (optional)}
}
\description{
Validate iteration history array
}
\keyword{internal}
