% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-validation.R
\name{.chck_ifca}
\alias{.chck_ifca}
\title{Validate IFCA-specific spax object}
\usage{
.chck_ifca(x, check_history = TRUE, name = "Input")
}
\description{
Validate IFCA-specific spax object
}
\keyword{internal}
