% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-spax_diag_convergence.R
\name{.chck_ifca_convergence}
\alias{.chck_ifca_convergence}
\title{Validate inputs for IFCA convergence checks}
\usage{
.chck_ifca_convergence(
  x,
  metrics,
  plot,
  by_facility = FALSE,
  relative_cutoff = NULL
)
}
\arguments{
\item{x}{spax object to validate}

\item{metrics}{character vector of requested metrics}

\item{plot}{logical for plot request}

\item{by_facility}{logical for facility-level breakdowns}

\item{relative_cutoff}{numeric between 0 and 1 for defining relative thresholds}
}
\value{
Invisible TRUE if validation passes
}
\description{
Validate inputs for IFCA convergence checks
}
\keyword{internal}
