% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-spax_diag_facil.R
\name{.chck_ifca_facilities}
\alias{.chck_ifca_facilities}
\title{Validate inputs for IFCA facility analysis}
\usage{
.chck_ifca_facilities(x, relative_cutoff = NULL, plot)
}
\arguments{
\item{x}{spax object to validate}

\item{relative_cutoff}{Numeric between 0 and 1 for defining relative thresholds}

\item{plot}{Logical for plot request}
}
\value{
Invisible TRUE if validation passes
}
\description{
Validate inputs for IFCA facility analysis
}
\keyword{internal}
