% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-validation.R
\name{.chck_in_range}
\alias{.chck_in_range}
\title{Check if value is in valid range}
\usage{
.chck_in_range(x, min = 0, max = 1, inclusive = TRUE, name = "Value")
}
\description{
Check if value is in valid range
}
\keyword{internal}
