% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-validation.R
\name{.chck_is_raster}
\alias{.chck_is_raster}
\title{Check if the input is a SpatRaster - just a fast-track to .chck_class}
\usage{
.chck_is_raster(x, name = "Input")
}
\description{
Check if the input is a SpatRaster - just a fast-track to .chck_class
}
\keyword{internal}
