% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-validation.R
\name{.chck_length}
\alias{.chck_length}
\title{Check if length matches expected}
\usage{
.chck_length(len_x, expected, name = "Input")
}
\description{
Check if length matches expected
}
\keyword{internal}
