% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-validation.R
\name{.chck_lengths_match}
\alias{.chck_lengths_match}
\title{Check if length of two inputs match}
\usage{
.chck_lengths_match(
  len_x,
  len_y,
  name_x = "First input",
  name_y = "Second input"
)
}
\description{
Check if length of two inputs match
}
\keyword{internal}
