% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-sample_pmf.R, R/09-sample_pmf.R
\name{.chck_n_convert_pmf}
\alias{.chck_n_convert_pmf}
\title{Check and convert input to PMF}
\usage{
.chck_n_convert_pmf(x)

.chck_n_convert_pmf(x)
}
\arguments{
\item{x}{SpatRaster input}
}
\value{
List with PMF raster and total population

List with PMF raster and total population
}
\description{
Check and convert input to PMF

Check and convert input to PMF
}
\keyword{internal}
