% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-validation.R
\name{.chck_no_na}
\alias{.chck_no_na}
\title{Check if values contain NAs when not allowed}
\usage{
.chck_no_na(x, name = "Input")
}
\description{
Check if values contain NAs when not allowed
}
\keyword{internal}
