% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-validation.R
\name{.chck_nonnegative_scalar}
\alias{.chck_nonnegative_scalar}
\title{Check if input is a nonnegative scalar}
\usage{
.chck_nonnegative_scalar(x, name = "Input")
}
\description{
Check if input is a nonnegative scalar
}
\keyword{internal}
