% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-validation.R
\name{.chck_numeric_matrix}
\alias{.chck_numeric_matrix}
\title{Check if input is a numeric matrix}
\usage{
.chck_numeric_matrix(x, name = "Input", finite = TRUE)
}
\description{
Check if input is a numeric matrix
}
\keyword{internal}
