% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-validation.R
\name{.chck_numeric_vector}
\alias{.chck_numeric_vector}
\title{Check if input is a numeric vector}
\usage{
.chck_numeric_vector(x, name = "Input", finite = TRUE)
}
\description{
Check if input is a numeric vector
}
\keyword{internal}
