% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-validations_for_pmf_and_likelihood.R
\name{.chck_pmf_validity}
\alias{.chck_pmf_validity}
\title{Validate PMF properties}
\usage{
.chck_pmf_validity(x, tolerance = 1e-10)
}
\arguments{
\item{x}{SpatRaster to check PMF properties}

\item{tolerance}{Numeric tolerance for sum to 1 check}
}
\value{
Invisible TRUE if valid, error otherwise
}
\description{
Validate PMF properties
}
\keyword{internal}
