% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-validation.R
\name{.chck_positive}
\alias{.chck_positive}
\title{Check if value is positive}
\usage{
.chck_positive(x, allow_zero = FALSE, name = "Input")
}
\description{
Check if value is positive
}
\keyword{internal}
