% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-validation.R
\name{.chck_positive_integer}
\alias{.chck_positive_integer}
\title{Check if input is a positive integer scalar}
\usage{
.chck_positive_integer(x, name = "Input")
}
\description{
Check if input is a positive integer scalar
}
\keyword{internal}
