% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-validation.R
\name{.chck_positive_scalar}
\alias{.chck_positive_scalar}
\title{Check if input is a positive scalar}
\usage{
.chck_positive_scalar(x, name = "Input")
}
\description{
Check if input is a positive scalar
}
\keyword{internal}
