% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-validation.R
\name{.chck_raster_alignment}
\alias{.chck_raster_alignment}
\title{Check if rasters have matching geometry}
\usage{
.chck_raster_alignment(x, y, name_x = "First raster", name_y = "Second raster")
}
\description{
Check if rasters have matching geometry
}
\keyword{internal}
