% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-spax_field.R
\name{.chck_spax_field}
\alias{.chck_spax_field}
\title{Validate common spax_field inputs}
\usage{
.chck_spax_field(data, domain, index = NULL, role = "unknown", meta = list())
}
\description{
Validate common spax_field inputs
}
\keyword{internal}
