% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-spread_weighted.R
\name{.chck_spread_weighted}
\alias{.chck_spread_weighted}
\title{Validate inputs for spread_weighted}
\usage{
.chck_spread_weighted(values, weights, value_cols = NULL)
}
\arguments{
\item{values}{Numeric vector, matrix or data.frame of values}

\item{weights}{Multi-layer SpatRaster of weights}

\item{value_cols}{Column names if values is data.frame}
}
\description{
Validate inputs for spread_weighted
}
\keyword{internal}
