% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-transform_demand.R
\name{.chck_transform_demand}
\alias{.chck_transform_demand}
\title{Validate inputs for demand transformation}
\usage{
.chck_transform_demand(x, value_col, pmf = NULL, template = NULL)
}
\description{
Validate inputs for demand transformation
}
\keyword{internal}
