% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-transform_pmf.R
\name{.chck_transform_pmf}
\alias{.chck_transform_pmf}
\title{Validate inputs for PMF transformation}
\usage{
.chck_transform_pmf(x, snap = FALSE)
}
\arguments{
\item{x}{SpatRaster to be validated}

\item{snap}{Logical; if TRUE skip validation}
}
\value{
Invisible TRUE if valid, error otherwise
}
\description{
Validate inputs for PMF transformation
}
\keyword{internal}
