% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-transform_pmf.R
\name{.chck_transform_pmf_raster}
\alias{.chck_transform_pmf_raster}
\title{Validate inputs for PMF transformation}
\usage{
.chck_transform_pmf_raster(x, snap = FALSE)
}
\description{
Validate inputs for PMF transformation
}
\keyword{internal}
