% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-validations_for_pmf_and_likelihood.R
\name{.chck_vector_input}
\alias{.chck_vector_input}
\title{Validate SpatVector input}
\usage{
.chck_vector_input(x, value_col = NULL, template = NULL, allow_null = FALSE)
}
\arguments{
\item{x}{SpatVector to validate}

\item{value_col}{Character name of value column}

\item{template}{SpatRaster template for rasterization}

\item{allow_null}{Logical; if TRUE, value_col can be NULL}
}
\value{
Invisible TRUE if valid, error otherwise
}
\description{
Validate SpatVector input
}
\keyword{internal}
