% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_pmf.R
\name{.check_and_convert_pmf}
\alias{.check_and_convert_pmf}
\title{Check and convert input to PMF}
\usage{
.check_and_convert_pmf(x)
}
\arguments{
\item{x}{SpatRaster input}
}
\value{
List with PMF raster and total population
}
\description{
Check and convert input to PMF
}
\keyword{internal}
