% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-ae_problem.R
\name{.compile_ae_map}
\alias{.compile_ae_map}
\title{Compile a model spec and substrate into problem closures}
\usage{
.compile_ae_map(spec, substrate)
}
\description{
Compile a model spec and substrate into problem closures
}
\keyword{internal}
