% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-spax_e2sfca.R
\name{.compute_access_core}
\alias{.compute_access_core}
\title{Core computation for accessibility calculation}
\usage{
.compute_access_core(
  demand,
  supply_values,
  demand_weights,
  access_weights,
  indicator_names = NULL
)
}
\arguments{
\item{demand}{SpatRaster of demand}

\item{supply_values}{Matrix of supply values}

\item{demand_weights}{SpatRaster of demand-side weights}

\item{access_weights}{SpatRaster of accessibility-side weights}

\item{indicator_names}{Character vector of names for output layers}
}
\value{
SpatRaster of accessibility scores
}
\description{
Core computation for accessibility calculation
}
\keyword{internal}
