% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_choice.R
\name{.compute_choice_core}
\alias{.compute_choice_core}
\title{Internal function for core choice computation}
\usage{
.compute_choice_core(weights, attractiveness, alpha, a0, snap)
}
\description{
Internal function for core choice computation
}
\keyword{internal}
