% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_pmf.R
\name{.compute_sample_sizes}
\alias{.compute_sample_sizes}
\title{Compute sample sizes for each iteration}
\usage{
.compute_sample_sizes(
  n = NULL,
  size = NULL,
  prob = NULL,
  method = "poisson",
  iterations = 1,
  ...
)
}
\arguments{
\item{n}{Fixed sample size}

\item{size}{Population size}

\item{prob}{Probability parameter}

\item{method}{Sampling method}

\item{iterations}{Number of iterations}

\item{...}{Additional arguments for custom methods}
}
\value{
Vector of sample sizes
}
\description{
Compute sample sizes for each iteration
}
\keyword{internal}
