% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-spax_field.R
\name{.create_spax_field}
\alias{.create_spax_field}
\title{Create a generic spax_field with validation}
\usage{
.create_spax_field(
  data,
  domain,
  index = NULL,
  role = "unknown",
  meta = list(),
  subclass = NULL,
  snap = FALSE
)
}
\description{
Create a generic spax_field with validation
}
\keyword{internal}
