% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-spax_field.R
\name{.create_spax_vector_field}
\alias{.create_spax_vector_field}
\title{Create a vector spax_field}
\usage{
.create_spax_vector_field(
  data,
  domain,
  index = NULL,
  role = "unknown",
  meta = list(),
  allow_positional = FALSE,
  snap = FALSE
)
}
\description{
Create a vector spax_field
}
\keyword{internal}
