% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-decay_derivatives.R
\name{.decay_log_jacobian}
\alias{.decay_log_jacobian}
\title{Derivative of log decay weight with respect to the decay parameter}
\usage{
.decay_log_jacobian(method, distance, param)
}
\description{
Derivative of log decay weight with respect to the decay parameter
}
\keyword{internal}
