% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-decay_functions.R
\name{.exponential_weights}
\alias{.exponential_weights}
\title{Exponential decay weight computation}
\usage{
.exponential_weights(distance, sigma, ...)
}
\arguments{
\item{distance}{Numeric vector of distances}

\item{sigma}{Rate parameter controlling steepness of decay}

\item{...}{Additional parameters (not used in exponential decay)}
}
\value{
Numeric vector of weights
}
\description{
Exponential decay weight computation
}
\keyword{internal}
