% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-compute_fca.R
\name{.fca_kernel_axis_ids}
\alias{.fca_kernel_axis_ids}
\title{Return the non-I axis and IDs for a kernel field}
\usage{
.fca_kernel_axis_ids(kernel)
}
\description{
Return the non-I axis and IDs for a kernel field
}
\keyword{internal}
