% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-compute_fca.R
\name{.fca_result_raster}
\alias{.fca_result_raster}
\title{Unwrap a compute_fca result field as a public SpatRaster}
\usage{
.fca_result_raster(field)
}
\description{
Unwrap a compute_fca result field as a public SpatRaster
}
\keyword{internal}
