% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-spax_field.R
\name{.field_has_semantic_layer_names}
\alias{.field_has_semantic_layer_names}
\title{Detect whether raster names look like terra-generated defaults}
\usage{
.field_has_semantic_layer_names(x)
}
\description{
Detect whether raster names look like terra-generated defaults
}
\keyword{internal}
