% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-spax_field.R
\name{.field_index_frame}
\alias{.field_index_frame}
\title{Backend-neutral index frame accessor}
\usage{
.field_index_frame(field)
}
\description{
Backend-neutral index frame accessor
}
\keyword{internal}
