% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-ae_problem.R
\name{.fit_problem_decay}
\alias{.fit_problem_decay}
\title{Fit one decay parameter for an AE problem by weighted SSE}
\usage{
.fit_problem_decay(
  problem,
  observed,
  init,
  lower,
  upper,
  lambda = 1,
  tol = 1e-08,
  max_iter = 1000,
  eta = 1,
  control = list(maxit = 25),
  check = FALSE
)
}
\description{
Fit one decay parameter for an AE problem by weighted SSE
}
\keyword{internal}
