% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-ae_problem.R
\name{.fit_target_shape_ok}
\alias{.fit_target_shape_ok}
\title{Does a predicted output have the same shape as the calibration target?}
\usage{
.fit_target_shape_ok(predicted, target)
}
\description{
Does a predicted output have the same shape as the calibration target?
}
\keyword{internal}
