% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-gather_functions.R
\name{.gather_weighted_core}
\alias{.gather_weighted_core}
\title{Core computation for gather_weighted}
\usage{
.gather_weighted_core(values, weights, na.rm = TRUE)
}
\description{
Core computation for gather_weighted
}
\keyword{internal}
