% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-decay_functions.R
\name{.gaussian_weights}
\alias{.gaussian_weights}
\title{Gaussian decay weight computation}
\usage{
.gaussian_weights(distance, sigma = 11.64884, ...)
}
\arguments{
\item{distance}{Numeric vector of distances}

\item{sigma}{Standard deviation parameter controlling spread}

\item{...}{Additional parameters (not used in gaussian decay)}
}
\value{
Numeric vector of weights
}
\description{
Gaussian decay weight computation
}
\keyword{internal}
