% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_pmf.R
\name{.generate_evolve_samples}
\alias{.generate_evolve_samples}
\title{Generate evolving spatial samples across iterations}
\usage{
.generate_evolve_samples(
  x,
  n_samples,
  iterations,
  evolve_prop = 0.1,
  replace_0 = TRUE,
  snap = FALSE
)
}
\description{
Generate evolving spatial samples across iterations
}
\keyword{internal}
