% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-spax_haae.R
\name{.haae_compact_plan}
\alias{.haae_compact_plan}
\title{Compile inputs for the compact HAAE map}
\usage{
.haae_compact_plan(
  demand,
  supply,
  kernel,
  kappa = 1,
  id_col = NULL,
  supply_cols = NULL
)
}
\description{
Compile inputs for the compact HAAE map
}
\keyword{internal}
