% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-spax_haae.R
\name{.haae_predict_decay}
\alias{.haae_predict_decay}
\title{Predict HAAE utilization for a decay parameter}
\usage{
.haae_predict_decay(
  theta,
  family,
  demand,
  supply,
  distance,
  kappa = 1,
  eps = 1e-08,
  a_min = 1e-06,
  a_max = Inf,
  lambda = 0.7,
  tol = 1e-08,
  max_iter = 1000,
  x0 = NULL,
  check = FALSE
)
}
\description{
Predict HAAE utilization for a decay parameter
}
\keyword{internal}
