% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-ae_problem.R
\name{.haae_spec}
\alias{.haae_spec}
\title{Construct an HAAE model spec}
\usage{
.haae_spec(
  family = c("gaussian", "exponential", "power"),
  kappa = 1,
  eps = 1e-08,
  a_min = 1e-06,
  a_max = Inf
)
}
\description{
Construct an HAAE model spec
}
\keyword{internal}
