% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-spax_diagnoses.R
\name{.help_calc_changes}
\alias{.help_calc_changes}
\title{Calculate changes between iterations}
\usage{
.help_calc_changes(x, type = "absolute", na.rm = TRUE)
}
\arguments{
\item{x}{Numeric matrix [iterations, units] or vector}

\item{type}{Character; "absolute" or "relative"}

\item{na.rm}{Logical; whether to remove NA values}
}
\value{
Matrix or vector of changes
}
\description{
Calculate changes between iterations
}
\keyword{internal}
