% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-calc_choice.R
\name{.help_calc_choice_core}
\alias{.help_calc_choice_core}
\title{Internal function for core choice computation}
\usage{
.help_calc_choice_core(weights, attractiveness, alpha, a0)
}
\description{
Internal function for core choice computation
}
\keyword{internal}
