% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-spax_diagnoses.R
\name{.help_calc_rolling_stats}
\alias{.help_calc_rolling_stats}
\title{Calculate rolling statistics}
\usage{
.help_calc_rolling_stats(x, window, stats = c("mean", "max", "sd"))
}
\arguments{
\item{x}{Numeric vector}

\item{window}{Integer window size}

\item{stats}{Character vector of statistics to compute}
}
\value{
List of rolling statistics
}
\description{
Calculate rolling statistics
}
\keyword{internal}
