% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-transform_demand.R
\name{.help_create_uniform_pmf}
\alias{.help_create_uniform_pmf}
\title{Create a uniform PMF from a template and vector}
\usage{
.help_create_uniform_pmf(x, template)
}
\arguments{
\item{x}{SpatVector containing areas}

\item{template}{SpatRaster template}
}
\value{
SpatRaster PMF with uniform weights within polygons
}
\description{
Create a uniform PMF from a template and vector
}
\keyword{internal}
