% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-decay_functions.R
\name{.help_decay_binary}
\alias{.help_decay_binary}
\title{Helper function for binary decay weight computation}
\usage{
.help_decay_binary(distance, sigma = 97.5, ...)
}
\arguments{
\item{distance}{Numeric vector of distances}

\item{sigma}{Distance threshold}

\item{...}{Additional parameters (not used in binary decay)}
}
\value{
Numeric vector of weights
}
\description{
Helper function for binary decay weight computation
}
\keyword{internal}
