% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-decay_functions.R
\name{.help_decay_gaussian}
\alias{.help_decay_gaussian}
\title{Helper function for Gaussian decay weight computation}
\usage{
.help_decay_gaussian(distance, sigma = 11.64884, ...)
}
\arguments{
\item{distance}{Numeric vector of distances}

\item{sigma}{Standard deviation parameter controlling spread}

\item{...}{Additional parameters (not used in gaussian decay)}
}
\value{
Numeric vector of weights
}
\description{
Helper function for Gaussian decay weight computation
}
\keyword{internal}
