% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-decay_functions.R
\name{.help_decay_power}
\alias{.help_decay_power}
\title{Helper function for power decay weight computation}
\usage{
.help_decay_power(distance, sigma, ...)
}
\arguments{
\item{distance}{Numeric vector of distances}

\item{sigma}{Exponent parameter controlling rate of decay}

\item{...}{Additional parameters (not used in power decay)}
}
\value{
Numeric vector of weights
}
\description{
Helper function for power decay weight computation
}
\keyword{internal}
