% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-sample_pmf.R, R/09-sample_pmf.R
\name{.help_est_new_sample}
\alias{.help_est_new_sample}
\title{Estimate new sample}
\usage{
.help_est_new_sample(input_vector, evolve_prop)

.help_est_new_sample(input_vector, evolve_prop)
}
\description{
Estimate new sample

Estimate new sample
}
\keyword{internal}
