% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-gather_functions.R
\name{.help_format_gather}
\alias{.help_format_gather}
\title{Format output for gather_weighted}
\usage{
.help_format_gather(weighted_sums)
}
\description{
Format output for gather_weighted
}
\keyword{internal}
