% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-transform_demand.R
\name{.help_transform_demand_core}
\alias{.help_transform_demand_core}
\title{Core computation for demand transformation}
\usage{
.help_transform_demand_core(
  x,
  value_col,
  pmf = NULL,
  template = NULL,
  density = FALSE
)
}
\description{
Core computation for demand transformation
}
\keyword{internal}
