% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-transform_likelihood.R
\name{.help_transform_likelihood_core}
\alias{.help_transform_likelihood_core}
\title{Core computation for likelihood transformation}
\usage{
.help_transform_likelihood_core(x, method = "mean", ref_value = NULL)
}
\description{
Core computation for likelihood transformation
}
\keyword{internal}
