% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-transform_pmf.R
\name{.help_transform_pmf_verify}
\alias{.help_transform_pmf_verify}
\title{Helper function to verify PMF properties}
\usage{
.help_transform_pmf_verify(pmf_raster)
}
\arguments{
\item{pmf_raster}{Computed PMF raster}
}
\value{
Invisible TRUE, with warning if precision issues detected
}
\description{
Helper function to verify PMF properties
}
\keyword{internal}
