% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-ae_problem.R
\name{.new_problem}
\alias{.new_problem}
\title{Create an AE inner problem object}
\usage{
.new_problem(model, substrate, state, theta, bind, metadata = list())
}
\description{
Create an AE inner problem object
}
\keyword{internal}
