% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-spax_field.R
\name{.new_spax_raster_field}
\alias{.new_spax_raster_field}
\title{Low-level raster spax_field constructor}
\usage{
.new_spax_raster_field(data, domain, index, role = "unknown", meta = list())
}
\description{
Low-level raster spax_field constructor
}
\keyword{internal}
