% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-spax_field.R
\name{.new_spax_vector_field}
\alias{.new_spax_vector_field}
\title{Low-level vector spax_field constructor}
\usage{
.new_spax_vector_field(
  data,
  domain,
  index = NULL,
  role = "unknown",
  meta = list()
)
}
\description{
Low-level vector spax_field constructor
}
\keyword{internal}
