% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-iterative_chain_diag_plot.R
\name{.plot_accuracy_comparison}
\alias{.plot_accuracy_comparison}
\title{Create prediction accuracy plots}
\usage{
.plot_accuracy_comparison(summary)
}
\description{
Create prediction accuracy plots
}
\keyword{internal}
