% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-iterative_chain_diag_plot.R
\name{.plot_convergence_rate}
\alias{.plot_convergence_rate}
\title{Create convergence rate plot}
\usage{
.plot_convergence_rate(changes, facilities)
}
\description{
Create convergence rate plot
}
\keyword{internal}
