% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-iterative_chain_diag_plot.R
\name{.plot_convergence_speed}
\alias{.plot_convergence_speed}
\title{Create iterations to convergence plot}
\usage{
.plot_convergence_speed(summary)
}
\description{
Create iterations to convergence plot
}
\keyword{internal}
