% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-iterative_chain_diag_plot.R
\name{.plot_convergence_trajectory}
\alias{.plot_convergence_trajectory}
\title{Create convergence trajectory plot}
\usage{
.plot_convergence_trajectory(data, facilities)
}
\description{
Create convergence trajectory plot
}
\keyword{internal}
