% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-spax_diag_facil.R
\name{.plot_facility_convergence}
\alias{.plot_facility_convergence}
\title{Plot facility convergence characteristics}
\usage{
.plot_facility_convergence(convergence)
}
\arguments{
\item{convergence}{Data frame of convergence metrics}
}
\value{
ggplot object
}
\description{
Plot facility convergence characteristics
}
\keyword{internal}
