% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-spax_diag_facil.R
\name{.plot_facility_patterns}
\alias{.plot_facility_patterns}
\title{Plot facility utilization patterns}
\usage{
.plot_facility_patterns(patterns)
}
\arguments{
\item{patterns}{List of pattern metrics from .calc_facility_patterns}
}
\value{
ggplot object
}
\description{
Plot facility utilization patterns
}
\keyword{internal}
