% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-iterative_chain_diag_plot.R
\name{.plot_metrics_summary}
\alias{.plot_metrics_summary}
\title{Create metrics summary plot}
\usage{
.plot_metrics_summary(metrics)
}
\description{
Create metrics summary plot
}
\keyword{internal}
