% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-iterative_chain_diag_plot.R
\name{.plot_oscillation_index}
\alias{.plot_oscillation_index}
\title{Create oscillation index plot}
\usage{
.plot_oscillation_index(summary)
}
\description{
Create oscillation index plot
}
\keyword{internal}
