% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-ae_objectives.R
\name{.poisson_gradient}
\alias{.poisson_gradient}
\title{Gradient of the Poisson NLL with respect to parameters}
\usage{
.poisson_gradient(predicted, observed, sensitivity, eps = 1e-09)
}
\description{
Gradient of the Poisson NLL with respect to parameters
}
\keyword{internal}
