% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decay_functions.R
\name{.power_weights}
\alias{.power_weights}
\title{Power decay weight computation}
\usage{
.power_weights(distance, sigma, ...)
}
\arguments{
\item{distance}{Numeric vector of distances}

\item{sigma}{Exponent parameter controlling rate of decay}

\item{...}{Additional parameters (not used in power decay)}
}
\value{
Numeric vector of weights
}
\description{
Power decay weight computation
}
\keyword{internal}
