% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-spax_field.R
\name{.prepare_raster_layer_index}
\alias{.prepare_raster_layer_index}
\title{Prepare a raster layer frame and re-key raster layers}
\usage{
.prepare_raster_layer_index(
  data,
  domain,
  frame = NULL,
  allow_positional = FALSE,
  meta = list()
)
}
\description{
Prepare a raster layer frame and re-key raster layers
}
\keyword{internal}
