% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-ae_problem.R
\name{.problem_available_surfaces}
\alias{.problem_available_surfaces}
\title{List model outputs that can be rewrapped onto the demand-cell surface}
\usage{
.problem_available_surfaces(problem, outputs)
}
\description{
List model outputs that can be rewrapped onto the demand-cell surface
}
\keyword{internal}
