% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-ae_problem.R
\name{.problem_map}
\alias{.problem_map}
\title{Evaluate a problem map at one state and theta}
\usage{
.problem_map(problem, x, theta)
}
\description{
Evaluate a problem map at one state and theta
}
\keyword{internal}
