% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-spax_field.R
\name{.rewrap_field}
\alias{.rewrap_field}
\title{Rebuild a field with new data but unchanged axis structure}
\usage{
.rewrap_field(
  field,
  data,
  role = .field_role(field),
  meta = .field_meta(field)
)
}
\description{
Rebuild a field with new data but unchanged axis structure
}
\keyword{internal}
