% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-spax_sae.R
\name{.sae_equilibrium}
\alias{.sae_equilibrium}
\title{Solve the SAE serviceability fixed point}
\usage{
.sae_equilibrium(
  plan,
  x0 = NULL,
  beta = 20,
  eps = 1e-08,
  lambda = 0.7,
  tol = 1e-08,
  max_iter = 1000,
  norm = c("max", "l2"),
  keep_history = TRUE,
  warn = TRUE,
  check = TRUE
)
}
\description{
Solve the SAE serviceability fixed point
}
\keyword{internal}
