% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-spax_sae.R
\name{.sae_fit_decay}
\alias{.sae_fit_decay}
\title{Fit one SAE decay parameter by weighted SSE}
\usage{
.sae_fit_decay(
  demand,
  supply,
  distance,
  observed,
  family = c("gaussian", "exponential", "power"),
  init,
  lower,
  upper,
  kappa = 1,
  beta = 20,
  eps = 1e-08,
  lambda = 0.7,
  tol = 1e-08,
  max_iter = 1000,
  eta = 1,
  control = list(maxit = 25)
)
}
\description{
Fit one SAE decay parameter by weighted SSE
}
\keyword{internal}
