% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-spax_sae.R
\name{.sae_predict_decay}
\alias{.sae_predict_decay}
\title{Predict SAE utilization for a decay parameter}
\usage{
.sae_predict_decay(
  theta,
  family,
  demand,
  supply,
  distance,
  kappa = 1,
  beta = 20,
  eps = 1e-08,
  lambda = 0.7,
  tol = 1e-08,
  max_iter = 1000,
  x0 = NULL,
  check = FALSE
)
}
\description{
Predict SAE utilization for a decay parameter
}
\keyword{internal}
