% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-spax_sae.R
\name{.sae_soft_cap}
\alias{.sae_soft_cap}
\title{Smooth cap at one with value zero at zero}
\usage{
.sae_soft_cap(x, beta = 20)
}
\description{
Smooth cap at one with value zero at zero
}
\keyword{internal}
