% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-spax_sae.R
\name{.sae_soft_cap_gradient}
\alias{.sae_soft_cap_gradient}
\title{Derivative of the SAE smooth cap}
\usage{
.sae_soft_cap_gradient(x, beta = 20)
}
\description{
Derivative of the SAE smooth cap
}
\keyword{internal}
