% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-ae_problem.R
\name{.sae_spec}
\alias{.sae_spec}
\title{Construct an SAE model spec}
\usage{
.sae_spec(
  family = c("gaussian", "exponential", "power"),
  kappa = 1,
  beta = 20,
  eps = 1e-08
)
}
\description{
Construct an SAE model spec
}
\keyword{internal}
