% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-sample_pmf.R, R/09-sample_pmf.R
\name{.sample_pmf_core_evlv}
\alias{.sample_pmf_core_evlv}
\title{Generate evolving spatial samples across iterations}
\usage{
.sample_pmf_core_evlv(
  x,
  n_samples,
  iterations,
  evolve_prop = 0.1,
  replace_0 = TRUE,
  snap = FALSE
)

.sample_pmf_core_evlv(
  x,
  n_samples,
  iterations,
  evolve_prop = 0.1,
  replace_0 = TRUE,
  snap = FALSE
)
}
\description{
Generate evolving spatial samples across iterations

Generate evolving spatial samples across iterations
}
\keyword{internal}
