% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-sample_pmf.R, R/09-sample_pmf.R
\name{.sample_pmf_core_indp}
\alias{.sample_pmf_core_indp}
\title{Generate spatial samples}
\usage{
.sample_pmf_core_indp(x, n_samples, iterations, replace_0 = TRUE, snap = FALSE)

.sample_pmf_core_indp(x, n_samples, iterations, replace_0 = TRUE, snap = FALSE)
}
\arguments{
\item{x}{PMF raster}

\item{n_samples}{Vector of sample sizes}

\item{iterations}{Number of iterations}

\item{replace_0}{Logical for NA replacement}

\item{snap}{Logical for snap mode}
}
\value{
Multi-layer SpatRaster of samples

Multi-layer SpatRaster of samples
}
\description{
Generate spatial samples

Generate spatial samples
}
\keyword{internal}
