% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-sample_pmf.R
\name{.sample_pmf_internal}
\alias{.sample_pmf_internal}
\title{Internal sampling function (moved from main body)}
\usage{
.sample_pmf_internal(
  x,
  n,
  size,
  prob,
  method,
  iterations,
  evolve_prop,
  replace_0,
  snap,
  ...
)
}
\description{
Internal sampling function (moved from main body)
}
\keyword{internal}
