% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-spax_field.R
\name{.spax_field_roles}
\alias{.spax_field_roles}
\title{Valid semantic roles for spax_field objects}
\usage{
.spax_field_roles()
}
\description{
Valid semantic roles for spax_field objects
}
\keyword{internal}
