% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_functions.R
\name{.spread_single}
\alias{.spread_single}
\title{Internal function for single value distribution}
\usage{
.spread_single(values, weights, full_output = TRUE)
}
\description{
Internal function for single value distribution
}
\keyword{internal}
