% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-spax_e2sfca.R
\name{.validate_facility_supply}
\alias{.validate_facility_supply}
\title{Core facility supply validation used by multiple functions}
\usage{
.validate_facility_supply(supply, id_col = NULL, supply_cols = NULL)
}
\description{
Core facility supply validation used by multiple functions
}
\keyword{internal}
