% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_pmf.R
\name{.validate_sample_pmf_inputs}
\alias{.validate_sample_pmf_inputs}
\title{Validate inputs for sample_pmf}
\usage{
.validate_sample_pmf_inputs(x, n, size, prob, method, iterations, snap)
}
\arguments{
\item{x}{SpatRaster input}

\item{n}{Sample size}

\item{size}{Population size}

\item{prob}{Probability parameter}

\item{method}{Sampling method}

\item{iterations}{Number of iterations}

\item{snap}{Logical for snap mode}
}
\description{
Validate inputs for sample_pmf
}
\keyword{internal}
