% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-ae_objectives.R
\name{.weighted_sse_loss}
\alias{.weighted_sse_loss}
\title{Weighted sum of squared errors}
\usage{
.weighted_sse_loss(predicted, observed, eta = 1)
}
\description{
Computes sum((predicted - observed)^2 / (observed + eta)).
}
\keyword{internal}
