% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-ae_solver.R
\name{equilibrium_residual}
\alias{equilibrium_residual}
\title{Fixed-point residual under F(x, theta) = x - T_theta(x)}
\usage{
equilibrium_residual(map, x, ..., check = TRUE)
}
\description{
Fixed-point residual under F(x, theta) = x - T_theta(x)
}
\keyword{internal}
