% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-ae_solver.R
\name{fd_jacobian}
\alias{fd_jacobian}
\title{Finite-difference Jacobian of a vector-valued function}
\usage{
fd_jacobian(
  fn,
  x,
  ...,
  eps = 1e-06,
  method = c("central", "forward"),
  check = TRUE
)
}
\description{
Finite-difference Jacobian of a vector-valued function
}
\keyword{internal}
