% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-ae_solver.R
\name{fd_jacobian_param}
\alias{fd_jacobian_param}
\title{Finite-difference parameter Jacobian dT/dtheta at fixed state x}
\usage{
fd_jacobian_param(
  map,
  theta,
  x,
  ...,
  eps = 1e-06,
  method = c("central", "forward"),
  check = TRUE
)
}
\description{
Finite-difference parameter Jacobian dT/dtheta at fixed state x
}
\keyword{internal}
