% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-ae_solver.R
\name{fd_jacobian_state}
\alias{fd_jacobian_state}
\title{Finite-difference state Jacobian dT/dx}
\usage{
fd_jacobian_state(
  map,
  x,
  ...,
  eps = 1e-06,
  method = c("central", "forward"),
  check = TRUE
)
}
\description{
Finite-difference state Jacobian dT/dx
}
\keyword{internal}
