% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-ae_coverage.R
\name{format.ae_coverage}
\alias{format.ae_coverage}
\title{Format a private coverage object}
\usage{
\method{format}{ae_coverage}(x, digits = 4, ...)
}
\description{
Format a private coverage object
}
\keyword{internal}
