% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain_access.R
\name{iterate_accessibility}
\alias{iterate_accessibility}
\title{Perform single iteration of accessibility calculation}
\usage{
iterate_accessibility(supply, supply_ratio, weights, demand)
}
\arguments{
\item{supply}{Original supply capacity vector}

\item{supply_ratio}{Current supply-to-demand ratios}

\item{weights}{Multi-layer SpatRaster of spatial weights}

\item{demand}{SpatRaster of demand locations}
}
\value{
List containing updated state information
}
\description{
Perform single iteration of accessibility calculation
}
\keyword{internal}
