% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-ae_atoms.R
\name{Math.spax_field}
\alias{Math.spax_field}
\title{Pointwise math transforms for spax_field objects}
\usage{
\method{Math}{spax_field}(x, ...)
}
\description{
Pointwise math transforms for spax_field objects
}
\keyword{internal}
