% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-ae_atoms.R
\name{Ops.spax_field}
\alias{Ops.spax_field}
\title{Pointwise arithmetic for same-domain spax_field objects}
\usage{
\method{Ops}{spax_field}(e1, e2)
}
\description{
Pointwise arithmetic for same-domain spax_field objects
}
\keyword{internal}
