% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-ae_methods.R
\name{plot.ae_problem_nfxp_fit}
\alias{plot.ae_problem_nfxp_fit}
\title{Plot AE problem calibration diagnostics}
\usage{
\method{plot}{ae_problem_nfxp_fit}(x, type = c("fit", "convergence", "state"), ...)
}
\description{
Plot AE problem calibration diagnostics
}
\keyword{internal}
