% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_assess_ifca.R
\name{plot.ifca_assessment}
\alias{plot.ifca_assessment}
\title{Plot method for ifca_assessment objects}
\usage{
plot.ifca_assessment(x, type = c("convergence", "accuracy", "both"), ...)
}
\arguments{
\item{x}{Object of class "ifca_assessment"}

\item{type}{Type of plots to display: "convergence", "accuracy", or "both"}

\item{...}{Additional arguments passed to plotting functions}
}
\description{
Plot method for ifca_assessment objects
}
