% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-ae_coverage.R
\name{print.ae_coverage}
\alias{print.ae_coverage}
\title{Print a private coverage object}
\usage{
\method{print}{ae_coverage}(x, digits = 4, ...)
}
\description{
Print a private coverage object
}
\keyword{internal}
