% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-ae_methods.R
\name{print.ae_equilibrium}
\alias{print.ae_equilibrium}
\title{Print an AE fixed-point solution}
\usage{
\method{print}{ae_equilibrium}(x, digits = 4, ...)
}
\description{
Print an AE fixed-point solution
}
\keyword{internal}
