% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-ae_methods.R
\name{print.ae_problem_nfxp_fit}
\alias{print.ae_problem_nfxp_fit}
\title{Print an AE problem calibration fit}
\usage{
\method{print}{ae_problem_nfxp_fit}(x, digits = 4, ...)
}
\description{
Print an AE problem calibration fit
}
\keyword{internal}
