% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-spax_diag_convergence.R
\name{print.ifca_convergence}
\alias{print.ifca_convergence}
\title{Print method for IFCA convergence diagnostics}
\usage{
\method{print}{ifca_convergence}(x, ...)
}
\arguments{
\item{x}{An ifca_convergence object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for IFCA convergence diagnostics
}
