% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-spax_field.R
\name{print.spax_field}
\alias{print.spax_field}
\title{Print a spax_field}
\usage{
\method{print}{spax_field}(x, ...)
}
\arguments{
\item{x}{A spax_field object}

\item{...}{Unused}
}
\value{
Invisibly returns `x`
}
\description{
Print a spax_field
}
