% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-ae_coverage.R
\name{print.summary.ae_coverage}
\alias{print.summary.ae_coverage}
\title{Print a private coverage summary}
\usage{
\method{print}{summary.ae_coverage}(x, ...)
}
\description{
Print a private coverage summary
}
\keyword{internal}
