% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-ae_methods.R
\name{print.summary.ae_problem_nfxp_fit}
\alias{print.summary.ae_problem_nfxp_fit}
\title{Print an AE problem calibration summary}
\usage{
\method{print}{summary.ae_problem_nfxp_fit}(x, digits = 4, ...)
}
\description{
Print an AE problem calibration summary
}
\keyword{internal}
