% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-ae_coverage.R
\name{summary.ae_coverage}
\alias{summary.ae_coverage}
\title{Summarize a private coverage object}
\usage{
\method{summary}{ae_coverage}(object, ...)
}
\description{
Summarize a private coverage object
}
\keyword{internal}
