% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-ae_methods.R
\name{summary.ae_problem_nfxp_fit}
\alias{summary.ae_problem_nfxp_fit}
\title{Summarize an AE problem calibration fit}
\usage{
\method{summary}{ae_problem_nfxp_fit}(object, ...)
}
\description{
Summarize an AE problem calibration fit
}
\keyword{internal}
