% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_chain.R
\name{validate_access_chain_inputs}
\alias{validate_access_chain_inputs}
\title{Validate inputs for accessibility chain algorithm}
\usage{
validate_access_chain_inputs(
  distance_raster,
  demand,
  supply,
  lambda,
  sigma,
  method
)
}
\description{
Validate inputs for accessibility chain algorithm
}
\keyword{internal}
