% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-ae_atoms.R
\name{.ae_nonI_axis}
\alias{.ae_nonI_axis}
\title{The single raster layer axis (errors on 0 or >1)}
\usage{
.ae_nonI_axis(field)
}
\description{
The single raster layer axis (errors on 0 or >1)
}
\keyword{internal}
