% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_assess_ifca.R
\name{.chck_assess_accuracy}
\alias{.chck_assess_accuracy}
\title{Validate inputs for accuracy assessment}
\usage{
.chck_assess_accuracy(predicted, observed, names = NULL)
}
\description{
Validate inputs for accuracy assessment
}
\keyword{internal}
