% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_assess_ifca.R
\name{.chck_assess_convergence}
\alias{.chck_assess_convergence}
\title{Validate inputs for convergence assessment}
\usage{
.chck_assess_convergence(history, names = NULL)
}
\description{
Validate inputs for convergence assessment
}
\keyword{internal}
