% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-spax_field.R
\name{.chck_spax_raster_field}
\alias{.chck_spax_raster_field}
\title{Validate raster spax_field inputs}
\usage{
.chck_spax_raster_field(
  data,
  domain,
  index = NULL,
  role = "unknown",
  meta = list()
)
}
\description{
This is the canonical finished-object validator for raster fields.
}
\keyword{internal}
