% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-ae_problem.R
\name{.fit_problem_nfxp}
\alias{.fit_problem_nfxp}
\title{Fit a private AE problem by nested fixed point}
\usage{
.fit_problem_nfxp(
  problem,
  observed,
  init,
  lower,
  upper,
  output = "utilization",
  loss = .weighted_sse_loss,
  loss_grad = .weighted_sse_gradient,
  loss_args = list(eta = 1),
  lambda = 1,
  tol = 1e-08,
  max_iter = 1000,
  norm = c("max", "l2"),
  control = list(maxit = 25),
  check = FALSE,
  penalty = 1e+12,
  gradient = FALSE,
  fd_eps = 1e-06
)
}
\description{
Fit a private AE problem by nested fixed point
}
\keyword{internal}
