% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-ae_problem.R
\name{.haae_problem}
\alias{.haae_problem}
\title{Construct a private HAAE problem}
\usage{
.haae_problem(
  demand,
  supply,
  distance,
  family = c("gaussian", "exponential", "power"),
  kappa = 1,
  eps = 1e-08,
  a_min = 1e-06,
  a_max = Inf,
  id_col = NULL,
  supply_cols = NULL
)
}
\description{
Construct a private HAAE problem
}
\keyword{internal}
