% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-sample_pmf.R
\name{.help_gen_sample_size}
\alias{.help_gen_sample_size}
\title{Compute sample sizes for each iteration}
\usage{
.help_gen_sample_size(
  n = NULL,
  size = NULL,
  prob = NULL,
  method = "poisson",
  iterations = 1,
  ...
)
}
\arguments{
\item{n}{Fixed sample size}

\item{size}{Population size}

\item{prob}{Probability parameter}

\item{method}{Sampling method}

\item{iterations}{Number of iterations}

\item{...}{Additional arguments for custom methods}
}
\value{
Vector of sample sizes
}
\description{
Compute sample sizes for each iteration
}
\keyword{internal}
