% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-spax_huff.R
\name{.huff_map}
\alias{.huff_map}
\title{Static Huff map T(a) = a (constant; reaches equilibrium in one pass)}
\usage{
.huff_map(a, plan, v0 = 0)
}
\description{
Static Huff map T(a) = a (constant; reaches equilibrium in one pass)
}
\keyword{internal}
