% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-ae_problem.R
\name{.interaction_substrate}
\alias{.interaction_substrate}
\title{Compile checked raster inputs into a compact interaction substrate}
\usage{
.interaction_substrate(
  demand,
  supply,
  distance,
  id_col = NULL,
  supply_cols = NULL
)
}
\description{
Compile checked raster inputs into a compact interaction substrate
}
\keyword{internal}
