% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-ae_problem.R
\name{.problem_outputs}
\alias{.problem_outputs}
\title{Evaluate rich outputs at one state and theta}
\usage{
.problem_outputs(problem, x, theta)
}
\description{
Evaluate rich outputs at one state and theta
}
\keyword{internal}
