% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-ae_problem.R
\name{.problem_outputs_at}
\alias{.problem_outputs_at}
\title{Evaluate rich outputs for a problem at arbitrary theta and state}
\usage{
.problem_outputs_at(problem, theta, state)
}
\description{
Evaluate rich outputs for a problem at arbitrary theta and state
}
\keyword{internal}
