% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-ae_problem.R
\name{.sae_problem}
\alias{.sae_problem}
\title{Construct a private SAE problem}
\usage{
.sae_problem(
  demand,
  supply,
  distance,
  family = c("gaussian", "exponential", "power"),
  kappa = 1,
  beta = 20,
  eps = 1e-08,
  id_col = NULL,
  supply_cols = NULL
)
}
\description{
Construct a private SAE problem
}
\keyword{internal}
