% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-ae_problem.R
\name{.solve_problem}
\alias{.solve_problem}
\title{Solve an AE problem at fixed theta}
\usage{
.solve_problem(
  problem,
  theta,
  x0 = NULL,
  lambda = 1,
  tol = 1e-08,
  max_iter = 1000,
  norm = c("max", "l2"),
  keep_history = TRUE,
  warn = TRUE,
  check = TRUE,
  diagnostics = TRUE
)
}
\description{
Solve an AE problem at fixed theta
}
\keyword{internal}
