% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-spread_functions.R
\name{.spread_core}
\alias{.spread_core}
\title{Internal function for single value distribution}
\usage{
.spread_core(values, weights, full_output = TRUE)
}
\description{
Internal function for single value distribution
}
\keyword{internal}
