% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-spread_weighted.R
\name{.spread_core}
\alias{.spread_core}
\title{Core spreading computation for single value vector}
\usage{
.spread_core(values, weights, full_output = FALSE)
}
\arguments{
\item{values}{Numeric vector of values}

\item{weights}{Multi-layer SpatRaster of weights}

\item{full_output}{Whether to return intermediate calculations}
}
\value{
SpatRaster or list depending on full_output
}
\description{
Core spreading computation for single value vector
}
\keyword{internal}
