% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_functions.R
\name{normalize_weights}
\alias{normalize_weights}
\title{Normalize Spatial Weights}
\usage{
normalize_weights(
  x,
  method = "standard",
  ref_value = NULL,
  a0 = 0,
  snap = FALSE,
  ...
)
}
\arguments{
\item{x}{Numeric vector, matrix, or SpatRaster of weights}

\item{method}{Character string specifying normalization method:

       - "standard": Divide by sum + a0 (default)

       - "semi": Normalize only if sum + a0 > 1

       - "reference": Normalize by reference value

       - "identity": Return unchanged

       - Or a custom function(x, ...)}

\item{ref_value}{Optional reference value for "reference" method}

\item{a0}{Non-negative numeric value representing outside option weight (default = 0) *Now working for only standard and semi method}

\item{snap}{Logical; if TRUE, avoid input validation. PS. snap is a performance optimization for internal use}

\item{...}{Additional arguments passed to custom normalization function}
}
\value{
Object of the same class as input containing normalized weights
}
\description{
Normalizes weights using different methods with support for outside options.
}
