% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-spax_diag_facil.R
\name{print.ifca_facilities}
\alias{print.ifca_facilities}
\title{Print method for IFCA facility diagnostics}
\usage{
\method{print}{ifca_facilities}(x, ...)
}
\arguments{
\item{x}{An ifca_facilities object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for IFCA facility diagnostics
}
