% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-spax_save_and_read.R
\name{read_spax}
\alias{read_spax}
\title{Read a spax object from disk}
\usage{
read_spax(file)
}
\arguments{
\item{file}{Character; path/name of saved spax files}
}
\value{
A spax object
}
\description{
Read a spax object from disk
}
