% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-spax_save_and_read.R
\name{save_spax}
\alias{save_spax}
\title{Save a spax accessibility analysis}
\usage{
save_spax(x, file, dir = FALSE, overwrite = FALSE, ...)
}
\arguments{
\item{x}{A spax object}

\item{file}{Character; path to save the object (without extension)}

\item{dir}{Logical; whether to create a directory with same name for files (default FALSE)}

\item{overwrite}{Logical; whether to overwrite existing files/directory (default FALSE)}

\item{...}{Additional arguments passed to terra::writeRaster()}
}
\value{
Invisibly returns the input object
}
\description{
Save a spax accessibility analysis
}
